Updated menu
It is now possible to go backwards in the LCD Menu. Battery voltages
have been removed and replaced with a low battery warning instead. You
will not be able to continue in the robot menu if the battery level is
critical. aka the low battery must be replaced to operate the robot
(robot will only be unoperable if the battery voltage is critical upon
start up, robot will continue to function if battery is critical if the
menu section has been passed).